+++ Examples/Algorithms/Digitization/src/DigitizationConfigurator.cpp
@@ -87,7 +87,8 @@ void ActsExamples::DigitizationConfigurator::operator()(
               unsigned int nBins = static_cast<unsigned int>(std::round(
                   (maxX - minX) / inputSegmentation.binningData()[0].step));
               outputSegmentation += Acts::BinUtility(
-                  nBins, minX, maxX, Acts::open, Acts::BinningValue::binX);
+                  nBins, static_cast<float>(minX), static_cast<float>(maxX),
+                  Acts::open, Acts::BinningValue::binX);
             }
             if (inputSegmentation.binningData()[0].binvalue ==
                     Acts::BinningValue::binY ||
@@ -100,7 +101,8 @@ void ActsExamples::DigitizationConfigurator::operator()(
                   std::round((maxY - minY) /
                              inputSegmentation.binningData()[accessBin].step));
               outputSegmentation += Acts::BinUtility(
-                  nBins, minY, maxY, Acts::open, Acts::BinningValue::binY);
+                  nBins, static_cast<float>(minY), static_cast<float>(maxY),
+                  Acts::open, Acts::BinningValue::binY);
             }
           } break;
 
@@ -114,7 +116,8 @@ void ActsExamples::DigitizationConfigurator::operator()(
               unsigned int nBins = static_cast<unsigned int>(std::round(
                   2 * maxX / inputSegmentation.binningData()[0].step));
               outputSegmentation += Acts::BinUtility(
-                  nBins, -maxX, maxX, Acts::open, Acts::BinningValue::binX);
+                  nBins, -static_cast<float>(maxX), static_cast<float>(maxX),
+                  Acts::open, Acts::BinningValue::binX);
             }
             if (inputSegmentation.binningData()[0].binvalue ==
                     Acts::BinningValue::binY ||
@@ -127,7 +130,8 @@ void ActsExamples::DigitizationConfigurator::operator()(
                   std::round((2 * maxY) /
                              inputSegmentation.binningData()[accessBin].step));
               outputSegmentation += Acts::BinUtility(
-                  nBins, -maxY, maxY, Acts::open, Acts::BinningValue::binY);
+                  nBins, -static_cast<float>(maxY), static_cast<float>(maxY),
+                  Acts::open, Acts::BinningValue::binY);
             }
           } break;
 
@@ -140,7 +144,8 @@ void ActsExamples::DigitizationConfigurator::operator()(
               unsigned int nBins = static_cast<unsigned int>(std::round(
                   (maxR - minR) / inputSegmentation.binningData()[0].step));
               outputSegmentation += Acts::BinUtility(
-                  nBins, minR, maxR, Acts::open, Acts::BinningValue::binR);
+                  nBins, static_cast<float>(minR), static_cast<float>(maxR),
+                  Acts::open, Acts::BinningValue::binR);
             }
             if (inputSegmentation.binningData()[0].binvalue ==
                     Acts::BinningValue::binPhi ||
@@ -156,9 +161,9 @@ void ActsExamples::DigitizationConfigurator::operator()(
               unsigned int nBins = static_cast<unsigned int>(
                   std::round((maxPhi - minPhi) /
                              inputSegmentation.binningData()[accessBin].step));
-              outputSegmentation +=
-                  Acts::BinUtility(nBins, minPhi, maxPhi, Acts::open,
-                                   Acts::BinningValue::binPhi);
+              outputSegmentation += Acts::BinUtility(
+                  nBins, static_cast<float>(minPhi), static_cast<float>(maxPhi),
+                  Acts::open, Acts::BinningValue::binPhi);
             }
 
           } break;
@@ -175,7 +180,8 @@ void ActsExamples::DigitizationConfigurator::operator()(
               unsigned int nBins = static_cast<unsigned int>(std::round(
                   (maxR - minR) / inputSegmentation.binningData()[0].step));
               outputSegmentation += Acts::BinUtility(
-                  nBins, minR, maxR, Acts::open, Acts::BinningValue::binR);
+                  nBins, static_cast<float>(minR), static_cast<float>(maxR),
+                  Acts::open, Acts::BinningValue::binR);
             }
             if (inputSegmentation.binningData()[0].binvalue ==
                     Acts::BinningValue::binPhi ||
@@ -195,7 +201,8 @@ void ActsExamples::DigitizationConfigurator::operator()(
               unsigned int nBins = static_cast<unsigned int>(std::round(
                   2 * alpha / inputSegmentation.binningData()[accessBin].step));
               outputSegmentation += Acts::BinUtility(
-                  nBins, averagePhi - alpha, averagePhi + alpha, Acts::open,
+                  nBins, static_cast<float>(averagePhi - alpha),
+                  static_cast<float>(averagePhi + alpha), Acts::open,
                   Acts::BinningValue::binPhi);
             }
 
@@ -209,7 +216,8 @@ void ActsExamples::DigitizationConfigurator::operator()(
               unsigned int nBins = static_cast<unsigned int>(std::round(
                   (maxR - minR) / inputSegmentation.binningData()[0].step));
               outputSegmentation += Acts::BinUtility(
-                  nBins, minR, maxR, Acts::open, Acts::BinningValue::binR);
+                  nBins, static_cast<float>(minR), static_cast<float>(maxR),
+                  Acts::open, Acts::BinningValue::binR);
             }
             if (inputSegmentation.binningData()[0].binvalue ==
                     Acts::BinningValue::binPhi ||
@@ -227,9 +235,9 @@ void ActsExamples::DigitizationConfigurator::operator()(
               unsigned int nBins = static_cast<unsigned int>(
                   std::round((maxPhi - minPhi) /
                              inputSegmentation.binningData()[accessBin].step));
-              outputSegmentation +=
-                  Acts::BinUtility(nBins, minPhi, maxPhi, Acts::open,
-                                   Acts::BinningValue::binPhi);
+              outputSegmentation += Acts::BinUtility(
+                  nBins, static_cast<float>(minPhi), static_cast<float>(maxPhi),
+                  Acts::open, Acts::BinningValue::binPhi);
             }
 
           } break;
